+++ R/impute.R
@@ -20,7 +20,7 @@
 #' vector to `f_i`; if `"row"`, passes each *row* separately; and if `"df"` passes the entire data set (data frame) to
 #' `f_i`. The function called by `f_i` should be compatible with the type of data passed to it.
 #' @param group_level A level of the framework to use for grouping indicators. This is only
-#' relevant if `impute_by = "row"`. In that case, indicators will be split into their groups at the
+#' relevant if `impute_by = "row"` or `"df"`. In that case, indicators will be split into their groups at the
 #' level specified by `group_level`, and imputation will be performed across rows of the group, rather
 #' than the whole data set. This can make more sense because indicators within a group are likely to be
 #' more similar.
@@ -164,7 +164,7 @@ Impute.purse <- function(x, dset, f_i = NULL, f_i_para = NULL, impute_by = "colu
 #' @param use_group Optional grouping variable name to pass to imputation function if this supports group
 #' imputation.
 #' @param group_level A level of the framework to use for grouping indicators. This is only
-#' relevant if `impute_by = "row"`. In that case, indicators will be split into their groups at the
+#' relevant if `impute_by = "row"` or `"df"`. In that case, indicators will be split into their groups at the
 #' level specified by `group_level`, and imputation will be performed across rows of the group, rather
 #' than the whole data set. This can make more sense because indicators within a group are likely to be
 #' more similar.
@@ -174,6 +174,8 @@ Impute.purse <- function(x, dset, f_i = NULL, f_i_para = NULL, impute_by = "colu
 #' frame.
 #' @param write_to Optional character string for naming the data set in the coin. Data will be written to
 #' `.$Data[[write_to]]`. Default is `write_to == "Imputed"`.
+#' @param disable Logical: if `TRUE` will disable imputation completely and write the unaltered data set. This option is mainly useful
+#' in sensitivity and uncertainty analysis (to test the effect of turning imputation on/off).
 #' @param ... arguments passed to or from other methods.
 #'
 #' @return An updated coin with imputed data set at `.$Data[[write_to]]`
@@ -190,11 +192,26 @@ Impute.purse <- function(x, dset, f_i = NULL, f_i_para = NULL, impute_by = "colu
 #'
 Impute.coin <- function(x, dset, f_i = NULL, f_i_para = NULL, impute_by = "column",
                         use_group = NULL, group_level = NULL, normalise_first = NULL, out2 = "coin",
-                        write_to = NULL, ...){
+                        write_to = NULL, disable = FALSE, ...){
 
   # WRITE LOG ---------------------------------------------------------------
 
   coin <- write_log(x, dont_write = "x")
+
+  # potentially skip all imputation
+  stopifnot(is.logical(disable))
+  if(disable){
+    idata <- get_dset(coin, dset = dset)
+    # output list
+    if(out2 == "df"){
+      return(idata)
+    } else {
+      if(is.null(write_to)){
+        write_to <- "Imputed"
+      }
+      return(write_dset(coin, idata, dset = write_to))
+    }
+  }
 
   # GET DSET, DEFAULTS ------------------------------------------------------
 
@@ -239,15 +256,15 @@ Impute.coin <- function(x, dset, f_i = NULL, f_i_para = NULL, impute_by = "colum
 
   } else {
 
-    if(group_level %nin% 1:x$Meta$maxlev){
-      stop("group_level is out of range: must be between 1 and max level.")
+    if(group_level %nin% 2:x$Meta$maxlev){
+      stop("group_level is out of range: must be between 2 and max level.")
     }
 
     # we split the data set based on its grouping at a given level
     # get lineage
     lin <- coin$Meta$Lineage
     # make sure lineage is in the same order as colnames of data
-    lin <- lin[match(colnames(iData_), lin$Indicator), ]
+    lin <- lin[match(colnames(iData_), lin[[1]]), ]
     # this is our factor variable for splitting cols
     f <- lin[[group_level]]
     # now split
@@ -424,6 +441,16 @@ Impute.data.frame <- function(x, f_i = NULL, f_i_para = NULL, impute_by = "colum
 
   if(impute_by == "df"){
 
+    # require that first arg of f_i is "x"
+    f_args <- formals(f_i)
+    if(is.null(f_args)){
+      stop("The function specified by f_i seems to have no input arguments!")
+    } else {
+      if(names(f_args)[1] != "x"){
+        stop("The first argument of f_i must be called 'x'.")
+      }
+    }
+
     # function args
     f_args <- list(x = x_n)
     if(!is.null(f_i_para)){
@@ -439,7 +466,7 @@ Impute.data.frame <- function(x, f_i = NULL, f_i_para = NULL, impute_by = "colum
     if(!is.data.frame(x_imp)){
       stop("Object returned by f_i is not a data frame.")
     }
-    if(dim(x_imp) != dim(x)){
+    if(!identical(dim(x_imp), dim(x))){
       stop("Object returned by f_i has different dimensions from x.")
     }
 
@@ -526,7 +553,7 @@ Impute.data.frame <- function(x, f_i = NULL, f_i_para = NULL, impute_by = "colum
 #' there are no `NA`s at all.
 #'
 #' @param x A numeric vector, possibly with `NA` values to be imputed.
-#' @param f_i A function that imputes missing values in a numeric vector. See descriotion and details.
+#' @param f_i A function that imputes missing values in a numeric vector. See description and details.
 #' @param f_i_para Optional further arguments to be passed to `f_i()`
 #' @param ... arguments passed to or from other methods.
 #'
@@ -548,6 +575,16 @@ Impute.numeric <- function(x, f_i = NULL, f_i_para = NULL, ...){
   # DEFAULTS ----------------------------------------------------------------
 
   f_i <- set_default(f_i, "i_mean")
+
+  # require that first arg of f_i is "x"
+  f_args <- formals(f_i)
+  if(is.null(f_args)){
+    stop("The function specified by f_i seems to have no input arguments!")
+  } else {
+    if(names(f_args)[1] != "x"){
+      stop("The first argument of f_i must be called 'x'.")
+    }
+  }
 
   # function args
   f_args <- list(x = x)
@@ -582,7 +619,7 @@ Impute.numeric <- function(x, f_i = NULL, f_i_para = NULL, ...){
   if(!is.numeric(xi)){
     stop("imputed vector is not numeric")
   }
-  if(any(xi[!nas] != x[!nas])){
+  if(!identical(xi[!nas], x[!nas])){
     stop("One or more non-NA values of x has changed as a result of imputation. Check the behaviour of the imputation function.")
   }
 
@@ -671,6 +708,8 @@ i_median <- function(x){
 #' @param x A numeric vector
 #' @param f A grouping variable, of the same length of `x`, that specifies the group that each value
 #' of `x` belongs to. This will be coerced to a factor.
+#' @param skip_f_na If `TRUE`, will work around any `NA`s in `f` (the corresponding values of `x` will be excluded from the imputation
+#' and returned unaltered). Else if `FALSE`, will cause an error.
 #'
 #' @return A numeric vector
 #' @export
@@ -680,15 +719,37 @@ i_median <- function(x){
 #' f <- c(rep("a", 6), rep("b", 6))
 #' i_mean_grp(x, f)
 #'
-i_mean_grp <- function(x, f){
+i_mean_grp <- function(x, f, skip_f_na = TRUE){
 
   stopifnot(is.numeric(x),
             length(x)==length(f))
 
+  # get any NAs in f
+  fna <- is.na(f)
+  if(sum(fna) == length(x)){
+    stop("f must have at least one non-NA value")
+  }
+
+  # extract x values with non-NA f values
+  if(skip_f_na){
+    x_use <- x[!fna]
+    f_use <- f[!fna]
+  } else {
+    if(any(fna)){
+      stop("NAs found in f. If skip_f_na = TRUE f cannot contain any NAs.")
+    }
+    x_use <- x
+    f_use <- f[!fna]
+  }
+
   # split by factors, apply func then unsplit
-  x_split <- split(x, f)
+  x_split <- split(x_use, f_use)
   x_split <- lapply(x_split, i_mean)
-  unsplit(x_split, f)
+  x_imp <- unsplit(x_split, f_use)
+
+  # reassemble and output
+  x[!fna] <- x_imp
+  x
 }
 
 #' Impute by group median