+++ tests/testthat/test-Impute.R
@@ -10,6 +10,12 @@ test_that("Impute.numeric", {
   expect_length(xi, length(x))
   expect_equal(xi[-3], x[-3])
 
+  # check for mis-specified function (first arg is not called x)
+  f_test <- function(y){y}
+  expect_error(
+    Impute(x, f_i = "f_test")
+  )
+
 })
 
 # Check data frame method: imputing using mean and median, by row and col
@@ -105,6 +111,10 @@ test_that("impute.coin", {
   # FINALLY the test
   expect_equal(xn$ConSpeed, as.numeric(xref_sc))
 
+  # also test no imputation
+  coin <- Impute(coin, dset = "Raw", f_i = "i_mean_grp", use_group = "Pop_group", write_to = "test1", disable = TRUE)
+  expect_identical(coin$Data$Raw, coin$Data$test1)
+
 
 })
 
@@ -128,5 +138,38 @@ test_that("impute.purse", {
   x22 <- pursei$coin$`2022`$Data$Imputed$Flights[1]
 
   expect_equal(x22, x21)
+
+})
+
+test_that("group_mean", {
+
+  # data plus groupings
+  x <- runif(20)
+  f <- sample(c("a","b"), 20, replace = TRUE)
+
+  # make some missing data
+  x[c(1, 5, 10)] <- NA
+
+  # impute
+  xi <- i_mean_grp(x, f)
+
+  # check
+  expect_equal(xi[1], mean(x[f==f[1]], na.rm = TRUE))
+  expect_equal(xi[5], mean(x[f==f[5]], na.rm = TRUE))
+  expect_equal(xi[10], mean(x[f==f[10]], na.rm = TRUE))
+
+  # now check case with missing groups
+  f[c(2,5)] <- NA
+  # impute
+  xi <- i_mean_grp(x, f)
+
+  # check - I exclude the NA group rows completely
+  x2 <- x[-c(2,5)]
+  f2 <- f[-c(2,5)]
+
+  # expect the first NA value to be equal to the mean of the group in the vector excluding NAs
+  expect_equal(xi[1], mean(x2[f2==f2[1]], na.rm = TRUE))
+  # expect second NA to be equal to NA still, because has NA group
+  expect_equal(xi[5], as.numeric(NA))
 
 })