+++ NEWS.md
@@ -1,3 +1,14 @@
+# COINr 1.1.1
+
+Lots of small fixes and feature extensions, including
+
+* More robust write to log, can deal with `COINr::func_Name` type calls
+* Catch spaces and numbers in iCodes
+* Tweaks to correlation tables
+* Reordering for correlation plots
+* Better input checks
+* Extra stats
+
 # COINr 1.1
 
 Some new functions and bug fixes: